==> 1.build_path.ipynb <==
import os
import pandas as pd
from tqdm import tqdm

base_dir = os.getcwd()

# projects_dir: Stores sequence files for each project
projects_dir = os.path.join(base_dir, 'projects')
# data_dir: Stores various configuration files
data_dir = os.path.join(base_dir, 'data')
# input_path_dir: Stores absolute paths of sequence files
input_path_dir = os.path.join(base_dir, 'input_path')

project_list = os.listdir(projects_dir)
project_list_dir = [os.path.join(projects_dir, item) for item in project_list]

config_temp_df = pd.DataFrame(columns=['project', 'seq_type', 'left_trim_len', 'trim_length'])
config_temp_df['project'] = project_list
seq_type = []

for project_dir_index, project_dir_item in tqdm(enumerate(project_list_dir)):
    print(project_dir_item)
    # list all fastq.gz files in each project directory
    sample_list = os.listdir(project_dir_item)
    sample_list = [sample for sample in sample_list if sample.endswith('.fastq.gz')]
    sample_list.sort()
    sample_name_list = [sample[:-9] for sample in sample_list]
    abs_path_list = [os.path.join(projects_dir, project_list[project_dir_index], sample) for sample in sample_list]
    # generate absolute paths for pair and single sample
    if sample_name_list[0][-2] == '_':
        seq_type.append('pair')
        sample_name_list = [sample_name[:-2] for sample_name in sample_name_list[::2]]
        input_path_df = pd.DataFrame(columns=['sample-id','forward-absolute-filepath','reverse-absolute-filepath'])
        input_path_df['sample-id'] = sample_name_list
        input_path_df['forward-absolute-filepath'] = abs_path_list[::2]
        input_path_df['reverse-absolute-filepath'] = abs_path_list[1::2]
    else:
        seq_type.append('single')
        input_path_df = pd.DataFrame(columns=['sample-id','absolute-filepath'])
        input_path_df['sample-id'] = sample_name_list
        input_path_df['absolute-filepath'] = abs_path_list
    input_path_df.to_csv(os.path.join(input_path_dir, project_list[project_dir_index] +'.txt'), sep='\t', index=False)
config_temp_df['seq_type'] = seq_type
config_temp_df.to_csv(os.path.join(data_dir, 'config_temp.csv'), index=False)
print('STEP 1 Done!')
