==> 2.prepare_for_denose.ipynb <==
# 2. prepare for denose
import os
import pandas as pd
import qiime2
from qiime2.plugins import demux, vsearch, quality_filter
from tqdm import tqdm

base_dir = os.getcwd()

# data_dir: Stores various configuration files
data_dir = os.path.join(base_dir, 'data')
# Stores intermediate results and final outputs.
results_dir = os.path.join(base_dir, 'results')
# input_path_dir: Stores absolute paths of sequence files
input_path_dir = os.path.join(base_dir, 'input_path')
# demux_seqs_dir: Stores raw sequences imported via QIIME 2.
demux_seqs_dir = os.path.join(results_dir, 'demux_seqs')
# per_sample_sequences_dir: Stores visualization files generated by QIIME 2 from raw sequences.
per_sample_sequences_demux_dir = os.path.join(results_dir, 'per_sample_sequences_demux')
# quality_filtered_sequences_dir: Stores sequences after quality filtering.
quality_filtered_sequences_dir = os.path.join(results_dir, 'quality_filtered_sequences')
# quality_filtered_sequences_demux_dir: Stores visualization files after quality filtering.
quality_filtered_sequences_demux_dir = os.path.join(results_dir, 'quality_filtered_sequences_demux')

config_df = pd.read_csv(os.path.join(data_dir, 'config_temp.csv'))

for index, row in tqdm(config_df.iterrows()):
    if row['seq_type'] == 'single':
        demux_seqs = qiime2.Artifact.import_data(type='SampleData[SequencesWithQuality]', 
                                                 view_type='SingleEndFastqManifestPhred33V2',
                                                 view= os.path.join(input_path_dir, row['project']+ '.txt'))
        
    else:
        demux_seqs_t = qiime2.Artifact.import_data(type='SampleData[PairedEndSequencesWithQuality]',
                                                 view_type='PairedEndFastqManifestPhred33V2',
                                                 view= os.path.join(input_path_dir, row['project']+ '.txt'))
        demux_seqs = vsearch.methods.merge_pairs(demux_seqs_t,
                                                 threads = 8).merged_sequences
    demux_seqs.save(os.path.join(demux_seqs_dir, row['project']))
    demux_summary = demux.visualizers.summarize(demux_seqs)
    demux_summary.visualization.save(os.path.join(per_sample_sequences_demux_dir, row['project']))

    quality_filtered_sequences, quality_filter_stats = quality_filter.methods.q_score(demux_seqs, min_quality = 25)
    quality_filtered_sequences.save(os.path.join(quality_filtered_sequences_dir, row['project']))
    quality_filtered_sequences_demux = demux.visualizers.summarize(quality_filtered_sequences)
    quality_filtered_sequences_demux.visualization.save(os.path.join(quality_filtered_sequences_demux_dir, row['project']))
    print(row['project'] + '\t' + str(index+1) + ' / ' + str(len(config_df)) + ' done')
print('STEP 2 Done!')
