==> 5.alpha_diversity.ipynb <==
import qiime2
import pandas as pd
import os
from qiime2.plugins import (feature_table, alignment, phylogeny, diversity)
from tqdm import tqdm

base_dir = os.getcwd()
results_dir = os.path.join(base_dir, 'results')
merged_dir = os.path.join(results_dir, 'merged')
data_dir = os.path.join(base_dir, 'data')

# merged_alpha_dir: Directory to save alpha diversity results
merged_alpha_dir = os.path.join(merged_dir, 'alpha')

def cal_alpha_diversity(sample_metadata_df, table_clean, rooted_tree):
    '''Calculate alpha diversity metrics and return a DataFrame with results.
    Args:
        sample_metadata_df (pd.DataFrame): DataFrame containing sample metadata.
        table_clean (qiime2.Artifact): Cleaned feature table artifact.
        rooted_tree (qiime2.Artifact): Rooted phylogenetic tree artifact.
    Returns:
        pd.DataFrame: DataFrame containing alpha diversity metrics.'''
    alpha_shannon = diversity.pipelines.alpha(table = table_clean, 
                                              metric = 'shannon').alpha_diversity.view(qiime2.Metadata).to_dataframe()
    alpha_evenness = diversity.pipelines.alpha(table = table_clean, 
                                                     metric = 'pielou_e').alpha_diversity.view(qiime2.Metadata).to_dataframe()
    alpha_observed_features = diversity.pipelines.alpha(table = table_clean, 
                                                        metric = 'observed_features').alpha_diversity.view(qiime2.Metadata).to_dataframe()
    alpha_faith_pd = diversity.pipelines.alpha_phylogenetic(table = table_clean, 
                                                            phylogeny = rooted_tree, 
                                                            metric='faith_pd').alpha_diversity.view(qiime2.Metadata).to_dataframe()

    alpha_df = pd.merge(sample_metadata_df[['ProjectID','HostName','BodySite', 'BodysiteClass']],
                        alpha_shannon, 
                        left_index=True, right_index=True, how='left')
    alpha_df = pd.merge(alpha_df, 
                        alpha_evenness, left_index=True, right_index=True, how='left')
    alpha_df = pd.merge(alpha_df, 
                        alpha_observed_features, left_index=True, right_index=True, how='left')
    alpha_df = pd.merge(alpha_df, 
                        alpha_faith_pd, left_index=True, right_index=True, how='left')
    return alpha_df

# Constructs a phylogenetic tree.
merged_rep_seqs = qiime2.Artifact.load(os.path.join(merged_dir, 'merged_rep_seqs.qza'))
aligned_rep_seqs = alignment.methods.mafft(merged_rep_seqs, n_threads = 16).alignment
masked_aligned_rep_seqs = alignment.methods.mask(aligned_rep_seqs).masked_alignment
masked_aligned_rep_seqs.save(os.path.join(merged_dir, 'masked_aligned_rep_seqs'))
masked_aligned_rep_seqs = qiime2.Artifact.load(os.path.join(merged_dir, 'masked_aligned_rep_seqs.qza'))
unrooted_tree = phylogeny.methods.fasttree(masked_aligned_rep_seqs, n_threads = 16).tree
unrooted_tree.save(os.path.join(merged_dir, 'unrooted_tree'))
rooted_tree = phylogeny.methods.midpoint_root(unrooted_tree).rooted_tree
rooted_tree.save(os.path.join(merged_dir, 'rooted_tree'))
# rooted_tree = qiime2.Artifact.load(os.path.join(merged_dir, 'rooted_tree.qza'))


sample_metadata = qiime2.Metadata.load(os.path.join(data_dir, 'sample_metadata.txt'))
merged_table_clean = qiime2.Artifact.load(os.path.join(merged_dir, 'merged_table_clean.qza'))

alpha_df = pd.DataFrame(
    columns=['SampleID','ProjectID', 'HostName', 'BodySite','shannon_entropy','pielou_evenness', 'observed_features', 'faith_pd']
)
alpha_df = cal_alpha_diversity(sample_metadata.to_dataframe(), merged_table_clean, rooted_tree)
alpha_df.reset_index(inplace=True)
alpha_df.to_csv(os.path.join(merged_alpha_dir, 'alpha_diversity.csv'), index=False)

print('STEP 5  Done!')

alpha_df
